ONNXToTorch: Update eltwise op
